+++ website/src/components/app-sidebar.tsx
@@ -200,7 +200,7 @@ export function ChatLeftSidebar({
         <div
             {...props}
             style={{
-                width: hideBrowser ? '100%' : '600px',
+                width: hideBrowser ? '100%' : '500px',
                 ...(hideBrowser && {
                   paddingLeft: '2px',
                   paddingRight: '2px',
@@ -208,7 +208,7 @@ export function ChatLeftSidebar({
             }}
             className='dark mx-auto bg-black h-full flex-col scheme-only-dark px-0 grid max-w-full min-h-full grid-rows-24 grid-cols-1 items-stretch gap-2'
         >
-            <div className='justify-between max-w-[900px] w-full mx-auto row-span-1 z-10 gap-2 flex px-6'>
+            <div className='justify-between max-w-[900px] w-full mx-auto row-span-1 z-10 gap-2 pr-2 flex'>
                 <TeamSwitcher className='grow ' sites={userSites} />
 
                 <div className='flex items-start gap-2'>
@@ -217,7 +217,7 @@ export function ChatLeftSidebar({
                 </div>
             </div>
 
-            <div className='p-0 grow relative overflow-y-auto  items-center overflow-x-hidden w-full row-span-23 flex flex-col '>
+            <div className='grow relative overflow-y-auto items-center overflow-x-hidden w-full row-span-23 flex flex-col '>
                 <Chat />
             </div>
         </div>